--- src/pages/ProductDetail/ProductDetailTop/DetailTop.jsx
@@ -1,9 +1,0 @@
-import { useEffect, useState } from 'react';
-import DetailTop from '../DetailTop.scss';
-
-const DetailTop = () => {
-  const [detailTop, setDetailTopData] = useState([]);
-  return <div>ProductList</div>;
-};
-
-export default DetailTop;
